+++ src/intervene/data.rs
@@ -4,6 +4,27 @@
 任务上下文保存的位置，在任务管理器 TaskManager ，在里面能找到一个数组 tasks ，其中的每一项都是一个任务控制块即 TaskControlBlock ，它负责保存一个任务的状态，而任务上下文 TaskContext 被保存在任务控制块中。在内核运行时我们会初始化 TaskManager 的全局实例 TASK_MANAGER ，因此所有任务上下文实际保存在在 TASK_MANAGER 中，从内存布局来看则是放在内核的全局数据 .data 段中。当我们将任务上下文保存完毕之后则转化为下图右侧的状态。当要从其他任务切换回来继续执行这个任务的时候，CPU 会读取同样的位置并从中恢复任务上下文。
 */
 
+use crate::cpu::DataReg;
+
+#[repr(C)]
+#[derive(Clone, Copy)]
+pub struct Data<DG: DataReg> {
+    pub data_reg: DG,
+    pub status: usize,
+    pub pc: usize,
+
+    pub kernel_info: KernelInfo,
+}
+
+impl<DG: DataReg> Data<DG> {
+    #[inline]
+    pub fn get_mut(frame_number: usize) -> &'static mut Self {
+        use crate::memory::Address;
+        let address = Address::address(frame_number);
+        unsafe{ &mut *(address as *mut Self) }
+    }
+}
+
 #[repr(C)]
 #[derive(Clone, Copy)]
 pub struct KernelInfo {
